Move prefabs and namespace
